core/standards/scripts/scheduler2/location_hash/419.html

<!doctype html>
<title>Location hash - and set hash many times with different values go back in history after timeout - check after timeout 2</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
//This test assumes that maximum history length per page is 50 (just like in chrome/firefox)
setup(function () {},{timeout:3000});
var t = async_test("", {timeout:2500});
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
});

var loops_number = 50;

window.onload = function () {
    setTimeout(step1, 100);  
};


function step1() {
    t.step(function() {
        var i = loops_number;
        do {
            location.hash = '#foo' + i;
            location.hash = '#bar' + i;
        } while (--i);
    });
    setTimeout(step2, 1500);
}

function step2() {
    t.step(function() {
        history.go(-49);
    });
    setTimeout(last_step,500);
}

function last_step() {
    t.step(function() {
        assert_equals( location.hash, '#foo' + 25);
        assert_equals( history.length, 50 );
    });
    t.done();
}
</script>
<div id="log"></div>
